refactor(business): 10.7 后端增加结算接口, 保存语音识别信息
- 移除原有的 BizFiletransController、BizFiletransBo、BizFiletransVo 和 BizFiletrans 实体类
- 新增 WebBizFiletransController 控制器，用于处理语音识别支付请求
- 更新 FiletransLangEnum枚举，增加新的语言选项- 重构 BizFiletransService接口，只保留 pay 方法- 优化 BizFiletransServiceImpl 实现类，专注于语音识别支付逻辑
- 新增 filetrans.http 文件，用于测试语音识别支付接口

+++ ruoyi-modules/ruoyi-business/src/main/java/org/dromara/business/domain/BizFiletrans.java
@@ -8,6 +8,7 @@ import lombok.EqualsAndHashCode;
 import org.dromara.common.tenant.core.TenantEntity;
 
 import java.io.Serial;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -43,12 +44,12 @@ public class BizFiletrans extends TenantEntity {
     /**
      * 音频文件时长|秒
      */
-    private Long second;
+    private Integer second;
 
     /**
      * 金额|元，second*单价
      */
-    private Long amount;
+    private BigDecimal amount;
 
     /**
      * 文件链接
